Main_Project/requirements.txt
streamlit
pandas
matplotlib
openpyxl
python-docx
PyPDF2
scikit-learn